Fix search result pane not resizable.
Because the search filters is inside (in the DOM) the search pane, the jquery resizable plugin was incorrectly adding events to that handle instead of it's own. Broken because we load filters on app load instead of search open to support search related feature.

Initializing the resizable plugin before initializing the search component fixes the issue.

[Delivers #65010704]