Prevent shake on map init
Was caused by the deferred loading of map.
1) Add vertex without location to graph after refresh
2) Load map
3) Map shakes

Because the vertexAdded event is deferred until map init (when user switches) it looks like the user just dragged the vertex onto the graph when the user switches to map. Change to always prevent shake events after catching up on the events

[Delivers #67938368]